+++ src/scene-ideas/d.cc
@@ -137,6 +137,6 @@ LetterD::LetterD()
     // Primitive state so that the draw call can issue the primitives we want.
     unsigned int curOffset(0);
     primVec_.push_back(PrimitiveState(GL_TRIANGLE_STRIP, 34, curOffset));
-    curOffset += (34 * sizeof(unsigned int));
+    curOffset += (34 * sizeof(unsigned short));
     primVec_.push_back(PrimitiveState(GL_LINE_STRIP, 34, curOffset));
 }
