+++ src/scene-ideas/o.cc
@@ -134,6 +134,6 @@ LetterO::LetterO()
     // Primitive state so that the draw call can issue the primitives we want.
     unsigned int curOffset(0);
     primVec_.push_back(PrimitiveState(GL_TRIANGLE_STRIP, 33, curOffset));
-    curOffset += (33 * sizeof(unsigned int));
+    curOffset += (33 * sizeof(unsigned short));
     primVec_.push_back(PrimitiveState(GL_LINE_STRIP, 33, curOffset));
 }
